requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.7
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.6
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
gunicorn==21.2.0
idna==3.4
npm==0.1.1
oauthlib==3.2.2
optional-django==0.1.0
packaging==23.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
social-auth-app-django==5.4.0
social-auth-core==4.5.1
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.6
whitenoise==6.6.0
psycopg2-binary==2.9.5
